+++ src/main/kotlin/app/Security.kt
@@ -1,18 +1,10 @@
-package com.dlfsystems
+package com.dlfsystems.app
 
 import com.auth0.jwt.JWT
 import com.auth0.jwt.algorithms.Algorithm
-import io.ktor.serialization.kotlinx.json.*
 import io.ktor.server.application.*
 import io.ktor.server.auth.*
 import io.ktor.server.auth.jwt.*
-import io.ktor.server.plugins.contentnegotiation.*
-import io.ktor.server.response.*
-import io.ktor.server.routing.*
-import io.ktor.server.websocket.*
-import io.ktor.websocket.*
-import java.time.Duration
-import kotlin.time.Duration.Companion.seconds
 
 fun Application.configureSecurity() {
     // Please read the jwt property from the config file if you are using EngineMain
