+++ src/test/kotlin/ApplicationTest.kt
@@ -1,5 +1,6 @@
 package com.dlfsystems
 
+import com.dlfsystems.app.module
 import io.ktor.client.request.*
 import io.ktor.http.*
 import io.ktor.server.testing.*
